lift/index.html

<html>
  <head>
    <link href="../tester/testharness.css" rel="stylesheet">
    <script src="../tester/testharness.js"></script>
  </head>
  <body>
    <style>
      body{
        margin: 0;
      }

      .wrapper{
        max-width: 1024px;
        margin: 0 auto;
        position: relative;
        height: 100%;
      }

      #liftCabine{
        border: 4px solid black;
        width: 150px;
        height: 200px;
        position: absolute;
        left: 50%;
        transform: translateX(-50%);
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
        bottom: 0;
      }

      #liftCabine.btn-0{
        bottom: -190px;
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
      }

      #liftCabine.btn-1{
        bottom: 0;
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
      }


      #liftCabine.btn-2{
        bottom: 50%;
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
      }


      #liftCabine.btn-3{
        bottom: calc(100% - 210px);
      }

      .buttons{
        width: 50px;
        position: absolute;
        right: 0;
        top: 50%;
        transform: translateY(-50%);
      }

      .buttons button{
        width: 50px;
        height: 50px;
        margin-bottom: 10px;
      }
    </style>
    <div class='wrapper'>
      <div id='liftCabine'></div>
      <div class='buttons'>
        <button id='btn-3'>3</button>
        <button id='btn-2'>2</button>
        <button id='btn-1'>1</button>
        <button id='btn-0'>-1</button>
      </div>
    </div>

<!--lift programme -->
  <script>
    var buttons = document.querySelectorAll('button');
    var liftCabine = document.querySelector('#liftCabine');

    buttons.forEach(function(item){
      item.addEventListener('click', function(e){
        liftCabine.className = e.target.id;
        console.log(e.target.id);
        console.log(liftCabine.className);
      });
    });
  </script>

<!--testing-->
  <script>
    // testing querySelectorAll
    var buttons = document.querySelectorAll('button');

    var testSelectorAll = async_test('testing querySelectorAll for floors-buttons');

    window.addEventListener('load', testSelectorAll.step_func(function(){
      var amount = 4;
      assert_equals(buttons.length, amount);
      this.done();
    }));

    //testing id of clicked button

    var testButtonsId = async_test('testing class of lift-cabine');

    testButtonsId.step(function(){
      buttons.forEach(function(item){
        item.addEventListener('click', testButtonsId.step_func(function(e){
          assert_equals(item.id, e.target.id);
          testButtonsId.done();
        }));
      });

      buttons.forEach(function(item){
        item.dispatchEvent(new Event('click'));
      });
    });

    // testing if id of button apply to lift cabine
    var liftCabine = document.querySelector('#liftCabine');

    var liftCabineTest = async_test('testing if lift cabine has class of clicked button');

    liftCabineTest.step(function(){
      buttons.forEach(function(item){
        item.addEventListener('click', liftCabineTest.step_func(function(e){
          assert_equals(liftCabine.className, e.target.id);
          liftCabineTest.done();
        }));
      });

      buttons.forEach(function(item){
        item.dispatchEvent(new Event('click'));
      });
    });
  </script>
  </body>
</html>
